pyproject.toml
[project]
name = "llm-judge"
version = "0.1.0"
description = "LLM evaluation pipeline: inference → autocheck → judge → compare"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "jsonschema>=4.0",
    "typer>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "tenacity>=8.0",
]

[project.scripts]
llm-judge = "llm_judge.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_judge"]
